+++ vignettes/link2GI6.R
@@ -1,4 +1,4 @@
-## ---- include = FALSE---------------------------------------------------------
+## ----include = FALSE----------------------------------------------------------
 knitr::opts_chunk$set(
   #fig.path = "https://raw.githubusercontent.com/r-spatial/link2GI/master/figures/",
   fig.align = "center",