+++ NEWS.md
@@ -1,132 +1,167 @@
 ## link2GI 0.5-4
+
 bugfixes:
-* issue #61 #62 #63
+
+*  #61 #62 #63
 
 
 ## link2GI 0.5-3
+
 bugfixes:
-* issue #59 , #58
+
+*  #59 , #58
 * fix rgrass7 dependencies
 
 ## link2GI 0.5-2
+
 new feature:
+
 * add return of vector data in OTB calls
 
 bugfixes:
-* issue #58
-* minor fixes in runOTB
 
-bugfixes:
-* prelim fix of issue #57
-* fix of issue #56
+*  #58
+* minor fixes in runOTB
+* prelim fix of  #57
+* fix of  #56
 * minor fixes
 
 new feature:
+
 * integration of stars and terra raster objects for linking GRASS
 * update of documentation
 
 ## link2GI 0.5-1
 
 bugfixes:
-* prelim fix of issue #57
-* fix of issue #56
+
+* prelim fix of  #57
+* fix of  #56
 * minor fixes
 
 new feature:
+
 * integration of stars and terra raster objects for linking GRASS
 * update of documentation
 
 ## link2GI 0.5-0
 
 bugfixes:
-* prelim fix of issue #56
+
+* prelim fix of  #56
 
 new feature:
+
 * changing to linkGRASS for both GRASS 7.x/8.x - keeping linkGRASS7for backwards compatibility 
 
 
 ## link2GI 0.4-7
 
 bugfixes:
-* issue #52
+
+*  #52
 
 ## link2GI 0.4-6
 
 bugfixes:
-* pre fix issue #52
+
+* pre fix  #52
 
 ## link2GI 0.4-5
 
 bugfixes:
-* fix issue #42, #44, #46, #48, #49
+
+* fix  #42, #44, #46, #48, #49
 
 new feature:
+
 * add installation vignette
 * update some documentation
 
 ## link2GI 0.4-4
 
 bugfixes:
-* fix issue #42
+
+* fix  #42
 
 ## link2GI 0.4-3
 
 bugfixes:
-* fix issue #40 
+
+* fix  #40 
 
 ## link2GI 0.4-2
 
 bugfixes:
-* fix issue #38
+
+* fix  #38
 
 
 ## link2GI 0.4-1
+
 new feature
+
 * add linkall
 * add direct reading and writing of vector data to GRASS sqlite
 
 bugfixes:
+
 * smaller bugfixes
 
 ## link2GI 0.4-0
+
 new feature
+
 * parse and return OTB API help for interactive use in R
-* add runOTB function according to # 29
+* add runOTB function according to #29
 
 bugfixes:
-* issue 19,21,22,26,29 fixed
+
+*  #19 #21 #22 #26 #29 fixed
 * first parsing of the new GRASS (since 7.8.x ) start file 
 
 ## link2GI 0.3-7
+
 new feature: 
+
 * improved implementation of parseOTB
 
 bugfixes:
-* several issues fixes
+
+* several fixes
 * examples and typos
 
 ## link2GI 0.3-6
+
 new feature: 
+
 * full re-implementation of linkGDAL returns now all gdal installations at a given search path
 
 bugfixes:
+
 * several fixes for Windows search and linking
 * examples and typos
 
 ## link2GI 0.3-5
+
 new feature: 
+
 * add manual build function for UNC pathes
 
 bugfixes:
+
 * remove search argument for PRZP
 * several fixes for Windows search and linking
 * Comments and typos
 
 ## link2GI 0.3-4
+
 new feature: 
+
 * add boolean variable if required API exists
 
 bugfixes:
+
 * fix of the quiet option
 * fix the search order
 * straighten GRASS problems with r.in.lidar
@@ -136,31 +171,41 @@ bugfixes:
 
 
 ## link2GI 0.3-3
+
 new feature: 
+
 * by default the latest version of each API is selected
 
 bugfixes:
+
 * OTB basic wrapper fixed for Windows
 * new examples
 * Typos
 
 ## link2GI 0.3-2
+
 new feature: 
+
 * ver_select can be used directly with the number of selection 
 * OTB basic wrapper
 
 bugfixes:
+
 * RSAGA related fixes
 * Typos
 
 ## link2GI 0.3-1
+
 new feature: 
+
 * add returnPaths argument for suppress all returns
 
 bugfixes:
+
 * windows RSAGA related fixes
 
 ## link2GI 0.3-0
+
 The 0.3-0  version is quite restructured to conform to both 
 common R conventions (e.g. no hidden globals) and to make the calls 
 more intuitive and one-line-like. some of the function follow a slightly 
@@ -168,29 +213,35 @@ different naming and argument setting. This was done to keep in line with furthe
 CLI linking functionality.
 
 new feature:
+
 * Linux support for multi-version selection
-* remove all global variables (re-opened issue #3)
+* remove all global variables (re-opened  #3)
 * homogenization of the link calls
 * wrap searchAPIOS functions with OS-agnostic findAPI functions
 * add a more robust return of all pathes and environmental variables as lists 
-* add a quiet option to supress all! console outputs (issue #4)
+* add a quiet option to supress all! console outputs ( #4)
   
 bugfixes:
+
 * remove tailing backslashes 
 * force compatibility to RSAGA and SAGA ver 3+
 
 ## link2GI 0.2-2
+
 new feature:
+
 * findGRASS a function that wraps for searchGRASSW and searchGRASSX
   
 ## link2GI 0.2-1
 
 bugfixes:
-* fix github issues #2 #3 #5 
+
+* fix github s #2 #3 #5 
   
 ## link2GI 0.2-0
 
 bugfixes:
+
 * fix several small bugs
 
 ## link2GI 0.1-0